Context refactor 2
