+++ src/PSRule/Pipeline/InvokePipelineBuilderBase.cs
@@ -131,42 +131,6 @@ internal abstract class InvokePipelineBuilderBase : PipelineBuilderBase, IInvoke
 
     protected override PipelineInputStream PrepareReader()
     {
-        if (!string.IsNullOrEmpty(Option.Input.ObjectPath))
-        {
-            AddVisitTargetObjectAction((sourceObject, next) =>
-            {
-                return PipelineReceiverActions.ReadObjectPath(sourceObject, next, Option.Input.ObjectPath, true);
-            });
-        }
-
-        if (Option.Input.Format == InputFormat.Yaml)
-        {
-            AddVisitTargetObjectAction((sourceObject, next) =>
-            {
-                return PipelineReceiverActions.ConvertFromYaml(sourceObject, next);
-            });
-        }
-        else if (Option.Input.Format == InputFormat.Json)
-        {
-            AddVisitTargetObjectAction((sourceObject, next) =>
-            {
-                return PipelineReceiverActions.ConvertFromJson(sourceObject, next);
-            });
-        }
-        else if (Option.Input.Format == InputFormat.Markdown)
-        {
-            AddVisitTargetObjectAction((sourceObject, next) =>
-            {
-                return PipelineReceiverActions.ConvertFromMarkdown(sourceObject, next);
-            });
-        }
-        else if (Option.Input.Format == InputFormat.PowerShellData)
-        {
-            AddVisitTargetObjectAction((sourceObject, next) =>
-            {
-                return PipelineReceiverActions.ConvertFromPowerShellData(sourceObject, next);
-            });
-        }
-        return new PipelineInputStream(VisitTargetObject, _InputPath, GetInputObjectSourceFilter(), Option);
+        return new PipelineInputStream(_InputPath, GetInputObjectSourceFilter(), Option);
     }
 }
